+++ cmd/next.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/EnderRealm/ticket/pkg/ticket"
+	"github.com/DustinReynoldsPE/ticket/pkg/ticket"
 	"github.com/spf13/cobra"
 )