+++ cmd/stats.go
@@ -5,7 +5,7 @@ import (
 	"math"
 	"time"
 
-	"github.com/EnderRealm/ticket/pkg/ticket"
+	"github.com/DustinReynoldsPE/ticket/pkg/ticket"
 	"github.com/spf13/cobra"
 )